Use absolute bucket numbers instead of a wrapping set of buckets
This fixes #45 and improves performance by calculcating the count
for subjects in redis by lua scripts without transferring the
values of the buckets

An additional maintenance call is necessary to cleanup old keys in
a subjects hash. If the subject is not accessed for a longer time
it will still be removed by the existing expiration.